remove RecipeBookComponent.initVisuals()
